requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dill==0.2.7.1
Flask==1.0.2
Flask-FileUpload==0.5.0
Flask-Login==0.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.4
google-auth==1.4.2
google-cloud==0.33.1
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.2
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.12.0
gunicorn==19.8.1
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
libsass==0.14.5
livereload==2.5.2
MarkupSafe==1.0
mysqlclient==1.3.12
oauth2client==3.0.0
peewee==3.4.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.2.post1
psutil==5.4.5
pyasn1==0.4.3
pyasn1-modules==0.2.1
PyMySQL==0.8.1
pyreadline==2.1
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
tornado==5.0.2
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
